+++ inc/admin/welcome-screen/sections/intro.php
@@ -11,13 +11,13 @@ $kouki = wp_get_theme( 'kouki' );
 <p style="font-size: 1.2em;margin: 0 0 2em;"><?php esc_html_e( 'Awesome! You\'ve decided to use Kouki as your theme.', 'kouki' ); ?></p>
 
 <div id="col-container">
-	<div id="col-right" style="padding-left: 30px;">		
+	<div id="col-right" style="padding-left: 30px;">
 		<h4 style="margin-top:0;"><?php esc_html_e( 'About this Theme', 'kouki' ); ?></h4>
 		<p><?php echo esc_attr( $kouki['Description'] ); ?></p>
-		<h4><?php esc_html_e( 'Recommended Plugins', 'kouki' ); ?></h4>		
+		<h4><?php esc_html_e( 'Recommended Plugins', 'kouki' ); ?></h4>
 		<p><?php echo sprintf( esc_html__( 'To get the most out of Kouki you should install and activate the following Plugin: %sOptions Framework%s', 'kouki' ), '<a href="https://wordpress.org/plugins/options-framework/">', '</a>' ); ?></p>
 		<h4><?php esc_html_e( 'Support', 'kouki' ); ?></h4>
-		<p><?php esc_html_e( 'As this is a free theme, support is limited to the basics. You can find me helping out in the designated support forum.', 'kouki' ); ?></p>
+		<p><?php echo sprintf( esc_html__( 'Please read the FAQs in the %sREADME.txt%s first before contacting support. As this is a free theme, support is limited to the basics. You can find me helping out in the designated support forum.', 'kouki' ), '<a href="https://github.com/felixdorner/kouki/blob/master/readme.txt">', '</a>' ); ?></p>
 		<p><a href="https://wordpress.org/support/theme/kouki" class="button"><?php esc_html_e( 'Support Forum', 'kouki' ); ?></a></p>
 		<h4><?php esc_html_e( 'Can I Contribute?', 'kouki' ); ?></h4>
 		<p><?php esc_html_e( 'Found a bug? Want to contribute a patch or create a new feature? GitHub is the place to go! Or would you like to translate Kouki in to your language?', 'kouki' ); ?></p>
@@ -32,7 +32,7 @@ $kouki = wp_get_theme( 'kouki' );
 </div>
 
 <p style="font-size: 1.2em; padding: 1.618em 0; margin: 1.618em 0 2.618em 0; border-top: 1px solid rgba(0,0,0,0.1); border-bottom: 1px solid rgba(0,0,0,0.1);">
-	<?php echo sprintf( esc_html__( 'Looking for more themes? Take a look at my website. %sWebsite%s', 'kouki' ), '<a href="http://felixdorner.de" class="button-primary" style="float: right;">', '</a>' ); ?>
+	<?php echo sprintf( esc_html__( 'Looking for more free themes? %sTake a look at my website%s', 'kouki' ), '<a href="http://felixdorner.de" class="button-primary" style="float: right;">', '</a>' ); ?>
 </p>
 
-<p><?php esc_html_e( 'Made with &#10084; in Berlin.', 'kouki' ); ?></p>
+<p><?php esc_html_e( 'Made with &#10084; in Berlin', 'kouki' ); ?> / <?php echo sprintf( esc_html__( '%sChangelog%s', 'kouki' ), '<a href="https://github.com/felixdorner/kouki/blob/master/changelog.md">', '</a>' ); ?></p>